Изменяет название, добавляет и экспортирует класс
