--- src/components/films-list.js
@@ -1,12 +1,0 @@
-const createFilmListTemplate = () => {
-  return (
-    `<section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container"></div>
-      </section>
-    </section>`
-  );
-};
-
-export {createFilmListTemplate};
